+++ modules/apps/portal-vulcan/portal-vulcan-api/bnd.bnd
@@ -4,4 +4,7 @@ Bundle-Version: 1.0.0
 Export-Package:\
 	com.liferay.portal.vulcan.context,\
 	com.liferay.portal.vulcan.dto,\
-	com.liferay.portal.vulcan.util
+	com.liferay.portal.vulcan.util,\
+	com.liferay.portal.vulcan.yaml,\
+	com.liferay.portal.vulcan.yaml.config,\
+	com.liferay.portal.vulcan.yaml.openapi
