+++ modules/apps/portal-vulcan/portal-vulcan-api/src/main/java/com/liferay/portal/vulcan/yaml/config/Security.java
@@ -12,21 +12,39 @@
  * details.
  */
 
-package com.liferay.portal.tools.rest.builder.internal.yaml.openapi;
+package com.liferay.portal.vulcan.yaml.config;
 
 /**
  * @author Peter Shin
  */
-public class Content {
+public class Security {
 
-	public Schema getSchema() {
-		return _schema;
+	public String getBasicAuth() {
+		return _basicAuth;
 	}
 
-	public void setSchema(Schema schema) {
-		_schema = schema;
+	public String getGuestAllowed() {
+		return _guestAllowed;
 	}
 
-	private Schema _schema;
+	public String getOAuth2() {
+		return _oAuth2;
+	}
+
+	public void setBasicAuth(String basicAuth) {
+		_basicAuth = basicAuth;
+	}
+
+	public void setGuestAllowed(String guestAllowed) {
+		_guestAllowed = guestAllowed;
+	}
+
+	public void setOAuth2(String oAuth2) {
+		_oAuth2 = oAuth2;
+	}
+
+	private String _basicAuth;
+	private String _guestAllowed;
+	private String _oAuth2;
 
 }
