+++ modules/apps/portal-vulcan/portal-vulcan-api/src/main/java/com/liferay/portal/vulcan/yaml/openapi/RequestBody.java
@@ -12,48 +12,23 @@
  * details.
  */
 
-package com.liferay.portal.tools.rest.builder.internal.yaml.openapi;
+package com.liferay.portal.vulcan.yaml.openapi;
+
+import java.util.Map;
 
 /**
  * @author Peter Shin
  */
-public class Parameter {
+public class RequestBody {
 
-	public String getIn() {
-		return _in;
+	public Map<String, Content> getContent() {
+		return _content;
 	}
 
-	public String getName() {
-		return _name;
+	public void setContent(Map<String, Content> content) {
+		_content = content;
 	}
 
-	public String getReference() {
-		return _reference;
-	}
-
-	public Schema getSchema() {
-		return _schema;
-	}
-
-	public void setIn(String in) {
-		_in = in;
-	}
-
-	public void setName(String name) {
-		_name = name;
-	}
-
-	public void setReference(String reference) {
-		_reference = reference;
-	}
-
-	public void setSchema(Schema schema) {
-		_schema = schema;
-	}
-
-	private String _in;
-	private String _name;
-	private String _reference;
-	private Schema _schema;
+	private Map<String, Content> _content;
 
 }
