+++ modules/apps/portal-vulcan/portal-vulcan-api/src/main/java/com/liferay/portal/vulcan/yaml/openapi/Response.java
@@ -12,10 +12,32 @@
  * details.
  */
 
-package com.liferay.portal.tools.rest.builder.internal.yaml.openapi;
+package com.liferay.portal.vulcan.yaml.openapi;
+
+import java.util.Map;
 
 /**
  * @author Peter Shin
  */
-public class Head extends Operation {
+public class Response {
+
+	public Map<String, Content> getContent() {
+		return _content;
+	}
+
+	public String getDescription() {
+		return _description;
+	}
+
+	public void setContent(Map<String, Content> content) {
+		_content = content;
+	}
+
+	public void setDescription(String description) {
+		_description = description;
+	}
+
+	private Map<String, Content> _content;
+	private String _description;
+
 }
