+++ modules/apps/portal-vulcan/portal-vulcan-api/src/main/java/com/liferay/portal/vulcan/yaml/openapi/Schema.java
@@ -12,15 +12,29 @@
  * details.
  */
 
-package com.liferay.portal.tools.rest.builder.internal.yaml.openapi;
+package com.liferay.portal.vulcan.yaml.openapi;
+
+import java.util.Map;
 
 /**
  * @author Peter Shin
  */
-public class Items {
+public class Schema {
+
+	public String getDescription() {
+		return _description;
+	}
 
 	public String getFormat() {
 		return _format;
+	}
+
+	public Items getItems() {
+		return _items;
+	}
+
+	public Map<String, Properties> getProperties() {
+		return _properties;
 	}
 
 	public String getReference() {
@@ -31,8 +45,20 @@ public class Items {
 		return _type;
 	}
 
+	public void setDescription(String description) {
+		_description = description;
+	}
+
 	public void setFormat(String format) {
 		_format = format;
+	}
+
+	public void setItems(Items items) {
+		_items = items;
+	}
+
+	public void setProperties(Map<String, Properties> properties) {
+		_properties = properties;
 	}
 
 	public void setReference(String reference) {
@@ -43,7 +69,10 @@ public class Items {
 		_type = type;
 	}
 
+	private String _description;
 	private String _format;
+	private Items _items;
+	private Map<String, Properties> _properties;
 	private String _reference;
 	private String _type;
