+++ modules/util/portal-tools-rest-builder/src/main/java/com/liferay/portal/tools/rest/builder/internal/freemarker/tool/JavaTool.java
@@ -19,25 +19,25 @@ import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.tools.rest.builder.internal.freemarker.tool.java.JavaParameter;
 import com.liferay.portal.tools.rest.builder.internal.freemarker.tool.java.JavaSignature;
 import com.liferay.portal.tools.rest.builder.internal.util.CamelCaseUtil;
-import com.liferay.portal.tools.rest.builder.internal.yaml.config.Application;
-import com.liferay.portal.tools.rest.builder.internal.yaml.config.ConfigYAML;
-import com.liferay.portal.tools.rest.builder.internal.yaml.openapi.Components;
-import com.liferay.portal.tools.rest.builder.internal.yaml.openapi.Content;
-import com.liferay.portal.tools.rest.builder.internal.yaml.openapi.Delete;
-import com.liferay.portal.tools.rest.builder.internal.yaml.openapi.Get;
-import com.liferay.portal.tools.rest.builder.internal.yaml.openapi.Head;
-import com.liferay.portal.tools.rest.builder.internal.yaml.openapi.Items;
-import com.liferay.portal.tools.rest.builder.internal.yaml.openapi.OpenAPIYAML;
-import com.liferay.portal.tools.rest.builder.internal.yaml.openapi.Operation;
-import com.liferay.portal.tools.rest.builder.internal.yaml.openapi.Options;
-import com.liferay.portal.tools.rest.builder.internal.yaml.openapi.Parameter;
-import com.liferay.portal.tools.rest.builder.internal.yaml.openapi.PathItem;
-import com.liferay.portal.tools.rest.builder.internal.yaml.openapi.Post;
-import com.liferay.portal.tools.rest.builder.internal.yaml.openapi.Properties;
-import com.liferay.portal.tools.rest.builder.internal.yaml.openapi.Put;
-import com.liferay.portal.tools.rest.builder.internal.yaml.openapi.RequestBody;
-import com.liferay.portal.tools.rest.builder.internal.yaml.openapi.Response;
-import com.liferay.portal.tools.rest.builder.internal.yaml.openapi.Schema;
+import com.liferay.portal.vulcan.yaml.config.Application;
+import com.liferay.portal.vulcan.yaml.config.ConfigYAML;
+import com.liferay.portal.vulcan.yaml.openapi.Components;
+import com.liferay.portal.vulcan.yaml.openapi.Content;
+import com.liferay.portal.vulcan.yaml.openapi.Delete;
+import com.liferay.portal.vulcan.yaml.openapi.Get;
+import com.liferay.portal.vulcan.yaml.openapi.Head;
+import com.liferay.portal.vulcan.yaml.openapi.Items;
+import com.liferay.portal.vulcan.yaml.openapi.OpenAPIYAML;
+import com.liferay.portal.vulcan.yaml.openapi.Operation;
+import com.liferay.portal.vulcan.yaml.openapi.Options;
+import com.liferay.portal.vulcan.yaml.openapi.Parameter;
+import com.liferay.portal.vulcan.yaml.openapi.PathItem;
+import com.liferay.portal.vulcan.yaml.openapi.Post;
+import com.liferay.portal.vulcan.yaml.openapi.Properties;
+import com.liferay.portal.vulcan.yaml.openapi.Put;
+import com.liferay.portal.vulcan.yaml.openapi.RequestBody;
+import com.liferay.portal.vulcan.yaml.openapi.Response;
+import com.liferay.portal.vulcan.yaml.openapi.Schema;
 
 import java.util.ArrayList;
 import java.util.Collections;
